restructured filebase
